Remove the image load transition to stop visual flickering issues
When considering things like memory vs disk browser cache, raced requests, devices with slow memory etc,
there doesn't seem to be a way of consistently preventing the animation from displaying when the image is already
cache, and playing the blur animation whenever it exceeds a certain threshold.

It always results in flickering, which is inconsistent and distracting. At least now the lack of animation means
the pop in has minimal flicker if the image is already cached, and when downloading the image the difference isn't that noticeable.